If given DAB_DOCKER_GID do not mount /etc/group
This mount is a bit sketchy but the docker group id is often needed but
not necessarily known by the end user. Latest docker for mac does not
allow mounting into /etc so this must be given via env var to skip
mounting.

Also includes some tweaks to formatting and mount unix docker socket to
a consistent location within dab.